+++ src/main/java/com/quaxt/mcc/optimizer/Optimizer.java
@@ -106,12 +106,12 @@ public class Optimizer {
                         aliasedVars.add(var1);
                     if (dstName.isStatic()) aliasedVars.add(dstName);
                 }
-                case CopyBitsFromOffset(ValIr v1, long _,int _, int _, VarIr dstName) -> {
+                case CopyBitsFromOffset(ValIr v1, int _,int _, int _, VarIr dstName) -> {
                     if (v1 instanceof VarIr var1 && var1.isStatic())
                         aliasedVars.add(var1);
                     if (dstName.isStatic()) aliasedVars.add(dstName);
                 }
-                case CopyBitsFromOffsetViaPointer(ValIr v1, long _,int _, int _, VarIr dstName) -> {
+                case CopyBitsFromOffsetViaPointer(ValIr v1, int _,int _, int _, VarIr dstName) -> {
                     if (v1 instanceof VarIr var1 && var1.isStatic())
                         aliasedVars.add(var1);
                     if (dstName.isStatic()) aliasedVars.add(dstName);
@@ -130,7 +130,7 @@ public class Optimizer {
                         aliasedVars.add(var1);
                     if (dstName.isStatic()) aliasedVars.add(dstName);
                 }
-                case CopyBitsToOffsetViaPointer(ValIr v1, VarIr dstName, long _, int _, int _) -> {
+                case CopyBitsToOffsetViaPointer(ValIr v1, VarIr dstName, int _, int _, int _) -> {
                     if (v1 instanceof VarIr var1 && var1.isStatic())
                         aliasedVars.add(var1);
                     if (dstName.isStatic()) aliasedVars.add(dstName);
